requirements.txt
tensorflow
numpy
Pillow
matplotlib
flask
flask-cors
